# api-docs.json
{"openapi":"3.0.1","info":{"title":"Order Service API","description":"Microservice for managing orders in the marketplace","contact":{"name":"IGDevX","url":"https://github.com/IGDevX"},"version":"1.0.0"},"servers":[{"url":"http://localhost:5003","description":"Generated server url"}],"paths":{"/orders":{"get":{"tags":["order-controller"],"operationId":"listOrders","responses":{"200":{"description":"OK","content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/OrderDetailDto"}}}}}}},"post":{"tags":["order-controller"],"operationId":"create","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateOrderRequest"}}},"required":true},"responses":{"200":{"description":"OK","content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateOrderResponse"}}}}}}},"/orders/{reference}/payment-intent":{"post":{"tags":["order-controller"],"operationId":"createPaymentIntent","parameters":[{"name":"reference","in":"path","required":true,"schema":{"type":"string"}}],"responses":{"200":{"description":"OK","content":{"application/json":{"schema":{"$ref":"#/components/schemas/PaymentIntentResponse"}}}}}}},"/internal/orders/{reference}/payment-confirmation":{"post":{"tags":["internal-controller"],"operationId":"paymentConfirmation","parameters":[{"name":"reference","in":"path","required":true,"schema":{"type":"string"}},{"name":"X-Service-Token","in":"header","required":true,"schema":{"type":"string"}}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/PaymentConfirmationRequest"}}},"required":true},"responses":{"200":{"description":"OK"}}}},"/orders/{id}":{"get":{"tags":["order-controller"],"operationId":"getOrder","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"OK","content":{"application/json":{"schema":{"$ref":"#/components/schemas/OrderDetailDto"}}}}}}},"/orders/producer/{producerId}":{"get":{"tags":["order-controller"],"operationId":"getOrdersByProducer","parameters":[{"name":"producerId","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"OK","content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/OrderDetailDto"}}}}}}}},"/orders/customer/{customerId}":{"get":{"tags":["order-controller"],"operationId":"getOrdersByCustomer","parameters":[{"name":"customerId","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"OK","content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/OrderDetailDto"}}}}}}}},"/health":{"get":{"tags":["root-controller"],"operationId":"health","responses":{"200":{"description":"OK","content":{"application/json":{"schema":{"type":"object","additionalProperties":{"type":"string"}}}}}}}},"/":{"get":{"tags":["root-controller"],"operationId":"root","responses":{"200":{"description":"OK","content":{"application/json":{"schema":{"type":"object","additionalProperties":{"type":"object"}}}}}}}}},"components":{"schemas":{"CreateOrderRequest":{"required":["customerId","items","producerId"],"type":"object","properties":{"producerId":{"type":"integer","format":"int64"},"customerId":{"type":"integer","format":"int64"},"items":{"type":"array","items":{"$ref":"#/components/schemas/OrderItemRequest"}},"idempotencyKey":{"type":"string"}}},"OrderItemRequest":{"required":["productId","quantity","unitPrice"],"type":"object","properties":{"productId":{"type":"integer","format":"int64"},"quantity":{"type":"integer","format":"int32"},"unitPrice":{"type":"number"}}},"CreateOrderResponse":{"type":"object","properties":{"id":{"type":"integer","format":"int64"},"reference":{"type":"string"}}},"PaymentIntentResponse":{"type":"object","properties":{"paymentIntentId":{"type":"string"},"status":{"type":"string"}}},"PaymentConfirmationRequest":{"required":["paymentIntentId","status"],"type":"object","properties":{"paymentIntentId":{"type":"string"},"status":{"type":"string"},"errorMessage":{"type":"string"}}},"OrderDetailDto":{"type":"object","properties":{"id":{"type":"integer","format":"int64"},"reference":{"type":"string"},"producerId":{"type":"integer","format":"int64"},"customerId":{"type":"integer","format":"int64"},"status":{"type":"string"},"totalAmount":{"type":"number"},"createdAt":{"type":"string","format":"date-time"},"items":{"type":"array","items":{"$ref":"#/components/schemas/OrderItemDetailDto"}}}},"OrderItemDetailDto":{"type":"object","properties":{"productId":{"type":"integer","format":"int64"},"quantity":{"type":"integer","format":"int32"},"unitPrice":{"type":"number"},"subtotal":{"type":"number"}}}}}}