# src/main/java/org/igdevx/spring_boot_order_microservice/dto/OrderItemDetailDto.java
package org.igdevx.spring_boot_order_microservice.dto;

import lombok.*;
import java.math.BigDecimal;

@Data
@NoArgsConstructor
@AllArgsConstructor
@Builder
public class OrderItemDetailDto {
    private Long productId;
    private Integer quantity;
    private BigDecimal unitPrice;
    private BigDecimal subtotal;
}